[backport] [CI] Require C++17 + CMake 3.18; Use CUDA 11.8 in CI (#8853)
