+++ R-package/CMakeLists.txt
@@ -31,7 +31,7 @@ if (USE_OPENMP)
 endif (USE_OPENMP)
 set_target_properties(
   xgboost-r PROPERTIES
-  CXX_STANDARD 14
+  CXX_STANDARD 17
   CXX_STANDARD_REQUIRED ON
   POSITION_INDEPENDENT_CODE ON)
