+++ tests/buildkite/build-jvm-packages-gpu.sh
@@ -14,5 +14,7 @@ else
 fi
 
 tests/ci_build/ci_build.sh jvm_gpu_build nvidia-docker \
-  --build-arg CUDA_VERSION_ARG=${CUDA_VERSION} tests/ci_build/build_jvm_packages.sh \
+  --build-arg CUDA_VERSION_ARG=${CUDA_VERSION} \
+  --build-arg NCCL_VERSION_ARG=${NCCL_VERSION} \
+  tests/ci_build/build_jvm_packages.sh \
   ${SPARK_VERSION} -Duse.cuda=ON ${arch_flag}
