+++ tests/buildkite/test-cpp-gpu.sh
@@ -20,4 +20,5 @@ tests/ci_build/ci_build.sh gpu nvidia-docker \
 # tests/ci_build/ci_build.sh rmm nvidia-docker \
 #   --build-arg CUDA_VERSION_ARG=$CUDA_VERSION \
 #   --build-arg RAPIDS_VERSION_ARG=$RAPIDS_VERSION bash -c \
+#   --build-arg NCCL_VERSION_ARG=$NCCL_VERSION bash -c \
 #   "source activate gpu_test && build/testxgboost --use-rmm-pool"
