+++ tests/ci_build/Dockerfile.gpu_build_r_centos7
@@ -36,8 +36,8 @@ RUN \
     bash Miniconda3.sh -b -p /opt/python && \
     /opt/python/bin/python -m pip install auditwheel awscli && \
     # CMake
-    wget -nv -nc https://cmake.org/files/v3.14/cmake-3.14.0-Linux-x86_64.sh --no-check-certificate && \
-    bash cmake-3.14.0-Linux-x86_64.sh --skip-license --prefix=/usr
+    wget -nv -nc https://cmake.org/files/v3.18/cmake-3.18.0-Linux-x86_64.sh --no-check-certificate && \
+    bash cmake-3.18.0-Linux-x86_64.sh --skip-license --prefix=/usr
 
 ENV GOSU_VERSION 1.10
